Error was a symbols relocation problem
set relocation-model to static on the rust linker/compiler options, the rust compiler incorrectly relocated call symboles to the functions location during the rust compilation phase. We wanted it to be relocated during the object file linking phase (so it has the correct memory location for the function), which now happens like it was supposed to.